moved encoding from internal so it can be used from elsewhere
