+++ cmd/prove.go
@@ -29,7 +29,7 @@ import (
 	groth16_bls381 "github.com/consensys/gnark/backend/bls381/groth16"
 	backend_bn256 "github.com/consensys/gnark/backend/bn256"
 	groth16_bn256 "github.com/consensys/gnark/backend/bn256/groth16"
-	"github.com/consensys/gnark/internal/utils/encoding/gob"
+	"github.com/consensys/gnark/encoding/gob"
 	"github.com/consensys/gurvy"
 	"github.com/spf13/cobra"
 )
