+++ internal/generators/backend/template/zkpschemes/tests_groth16.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 
 
-	"github.com/consensys/gnark/internal/utils/encoding/gob"
+	"github.com/consensys/gnark/encoding/gob"
 	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gurvy"
